requirements.txt
absl-py==2.1.0
accelerate==0.31.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
alabaster==0.7.13
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.4
contexttimer==0.3.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.14.0
distlib==0.3.8
docutils==0.20.1
easydict==1.9
einops==0.8.0
entrypoints==0.4
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fairscale==0.4.4
fastjsonschema==2.20.0
filelock==3.14.0
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.5.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==1.0.0
grpcio==1.63.0
h5py==3.11.0
huggingface-hub==0.23.4
identify==2.5.36
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_packaging==0.12.3
jupyter_server==2.14.1
jupyter_server_fileid==0.9.2
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaggle==1.6.14
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.8
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nbsphinx==0.8.7
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.13.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
openpyxl==3.1.5
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandoc==2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.22.0
plumbum==1.8.3
ply==3.11
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
pymilvus==2.4.4
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.4
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
salesforce-lavis==1.0.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.3
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==7.1.2
sphinx-autodoc-typehints==2.0.1
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.34.0
sympy==1.12
tenacity==8.3.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.4.12
tinycss2==1.3.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.26.1
triton==2.3.0
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.26.2
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==24.6.0
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.18.2
# python==3.8.0
